assistant/requirements.txt
fastapi==0.129.2
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
httpx==0.28.1
mcp==1.26.0
fastmcp==2.14.5
openai==2.28.0
google-genai==1.66.0
pydantic==2.12.5
pydantic-settings==2.13.0
openpyxl==3.1.2
pytest==8.3.5
pytest-asyncio==0.24.0
anyio[trio]==4.8.0
